fix: Arreglado error 400 Se ha arreglado el error 400 que se producia al hacer login con un usuario recien creado. Issue #161
