+++ src/main/java/com/nailing/app/usuario/UsuarioService.java
@@ -15,25 +15,7 @@ import com.nailing.app.centro.CentroService;
 public class UsuarioService {
 
     @Autowired
-    private UsuarioRepository usuarioRepository;
-    //añadir-actualizar usuario
-    public Usuario addUsuario(Map<String,String> map) {
-        String user = map.get("user");
-        List<Usuario> usuarios = this.findAll();
-        for(Usuario u: usuarios){
-            if(u.getUsuario().equals(user)) throw new IllegalArgumentException();
-        }
-        String contrasenya = map.get("password");
-        String email = map.get("email");
-        String telefono = map.get("telefono");
-        if(user == null || contrasenya == null|| email == null|| telefono == null){
-            throw new IllegalArgumentException();
-        }
-
-        Usuario usuario = new Usuario(user,contrasenya,email,telefono,"USER");
-
-        return usuarioRepository.save(usuario);
-    }
+    private UsuarioRepository usuarioRepository;  
     
     //encontrar usuario por su id
     public Optional<Usuario> findById(Long id) {
